update/ rename component to FileWrapper
