+++ modules/UserProfile/index.tsx
@@ -2,13 +2,13 @@ import React, { useState } from "react";
 import Image from "next/image";
 import Button from "@/common/components/GeneralButton";
 import ModalWrapper from "@/common/components/ModalWrapper";
-import WrapperFile from "@/common/components/WrapperFile";
+import FileWrapper from "@/common/components/FileWrapper";
 import selectActiveComponent from "@/common/helpers/selectActiveComponent";
 import checkIsFollowed from "@/common/helpers/checkIsFollowed";
 import TagBlock from "./TagBlock";
 import GroupsList from "./GroupsList";
 import Comments from "./Comments";
-import { ReturnComponentType } from "@/common/components/WrapperFile/data";
+import { ReturnComponentType } from "@/common/components/FileWrapper/data";
 import { userData, tabSet, ActiveTabType } from "./data";
 
 export default function UserProfile() {
@@ -96,13 +96,13 @@ export default function UserProfile() {
           </ModalWrapper>
         </section>
 
-        <WrapperFile
+        <FileWrapper
           tabSet={tabSet}
           activeTab={activeTab}
           setActiveTab={setActiveTab}
         >
           {selectActiveComponent(activeTab, returnComponent)}
-        </WrapperFile>
+        </FileWrapper>
       </div>
     </div>
   );
